-added a new video
