+++ .gitignore
@@ -1,0 +1,2 @@
+#Ignoring jetbrains IDE files
+.idea
