+++ WebApp/index.html
@@ -4,10 +4,10 @@
 	<head>
         <link rel="stylesheet" href="styles/index.css">
         <link rel="stylesheet" href="styles/video.css">
-        <script src="video.js" type="module"></script>
+        <script src="scripts/video.js"></script>
     </head>
     <body onload="init()">
-        <div id="videooverview"/>
+    <div id="videooverview"></div>
         <button id="backtovideos" onclick="showOverviewHideVideoplayer()" style="display: none">Back to videos</button>
         <div id="video" style="display: none;">
             <video id="videoplayer" width="800" height="450" controls>
