+++ WebApp/scripts/video.js
@@ -1,9 +1,9 @@
 "use strict";
 
-import {loadVideos} from "./videooverview";
-
 const forbidden=['<','>'];
 
+//*************************************Classes**************************************
+//NOT USED YET
 class Comment{
     constructor(author,message){
         var date = new Date();
@@ -14,6 +14,8 @@ class Comment{
 }
 
 //*************************************Helpers**************************************
+
+//creats XMLHttp- or ActiveX-Request depending on browser support
 function createAjaxRequest(){
     var request;
     if(window.XMLHttpRequest){
@@ -24,6 +26,7 @@ function createAjaxRequest(){
     return request;
 }
 
+//Checks if string contains illegal characters which are defined in forbidden
 function isInputLegal(strIn){
     var inputLength = strIn.length;
     var forbiddenLength = forbidden.length;
@@ -36,10 +39,31 @@ function isInputLegal(strIn){
     }
     return true;
 }
-//*************************************Initializer************************************
-function init(){
-    loadVideos()
-    initVideoPlayer()
+//*************************************Initializers************************************
+
+//Creates an entry on the video-overview for every video listed in the videos.json
+function initVideoOverview(){
+    var request = createAjaxRequest();
+    request.onreadystatechange = function(){
+        if(4 == this.readyState && 200 == this.status){
+            var videos = JSON.parse(this.responseText);
+            var videoOverview = document.getElementById("videooverview");
+
+            for(video of videos){
+                var header5 = document.createElement("h5");
+                var header7 = document.createElement("h7");
+                var ahref = document.createElement("a");
+                ahref.href = "javascript:showVideoPlayerHideOverview(" + "'" + JSON.stringify(video)+ "'" + ")";
+                ahref.innerHTML = video.name;
+                header5.appendChild(ahref);
+                header7.innerHTML = video.duration;
+                videoOverview.appendChild(header5);
+                videoOverview.appendChild(header7);
+            }
+        }
+    }
+    request.open("GET","./videos.json",true);
+    request.send();
 }
 
 function initVideoPlayer(){
@@ -48,9 +72,14 @@ function initVideoPlayer(){
     document.getElementById("videotitle").innerHTML = videoName;
 }
 
-//*************************************Others*****************************************
+//*************************************HTML-called-functions***********************************
+//Is called on page load, calls all initializers
+function init(){
+    initVideoOverview()
+    initVideoPlayer()
+}
 
-
+//Shows the video-player and hides the video-list
 function showVideoPlayerHideOverview(videoStr){
     var vidPlayer = document.getElementById("video");
     if(vidPlayer.style.display == "none") {
@@ -61,13 +90,14 @@ function showVideoPlayerHideOverview(videoStr){
         var buttonBackToVideos = document.getElementById("backtovideos");
 
         vidOverview.style.display = "none";
-        vidPlayer.style.display = "block";
         videoSrc.src = video.src;
         videoTitle.innerHTML = video.name;
         buttonBackToVideos.style.display = "block";
+        vidPlayer.style.display = "block";
+
     }
 }
-
+//Shows the video-list and hides the video-player
 function showOverviewHideVideoplayer(){
     var vidOverview = document.getElementById("videooverview");
     if(vidOverview.style.display == "none"){
@@ -79,7 +109,7 @@ function showOverviewHideVideoplayer(){
     }
 }
 
-//Erstellt einen neuen Kommentar
+//Creats a new comment
 function submitComment(){
     var newComment = document.createElement("div");
     var authorInput = document.getElementById("inputname");
@@ -91,18 +121,7 @@ function submitComment(){
     newComment.setAttribute("class","comment");
     newComment.innerHTML = "<h3>" + authorInput.value + " - " + date.toDateString() + " " + date.toTimeString() +  "</h3>" + messageInput.value;
     document.getElementById("commentsection").appendChild(newComment);
-    
+
     authorInput.value = "";
     messageInput.value = "";
 }
-
-export {createAjaxRequest,init,initVideoPlayer,showOverviewHideVideoplayer,showVideoPlayerHideOverview,submitComment}
-
-
-
-
-
-
-
-
-
