+++ WebApp/videos.json
@@ -1,3 +1,4 @@
-{"videos":[
-  { "id":"v00", "src":"media/videos/funnyvid0.mp4", "name": "funnyVid00.mp4", "duration": "0:03"}
-]}
+[
+  { "id":"v00", "src":"media/videos/funnyvid0.mp4", "name": "funnyVid00.mp4", "duration": "0:03"},
+  { "id":"v01", "src":"media/videos/wendler_egal.mp4", "name": "wendler_egal.mp4", "duration": "0:03"}
+]
